requirements.txt
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
python-frontmatter==1.0.0
python-slugify==7.0.0
PyYAML==6.0
text-unidecode==1.3
tomli==2.0.1
